rename multiple-students-form -> students-form
